requirements.txt
PyPDF2==3.0.1
pdfplumber==0.10.3
pandas 
python-dotenv 
unidecode
openpyxl